Add function to convert []net.IP to IP ranges
